requirements.txt
Django==1.11.2
Django-pytest==3.1.2
Pillow==4.2.0
pyjwt==1.5.2
Pyrebase==3.0.27
